requirements.txt
Flask
google-generativeai
requests
numpy
pandas